Rename rounders.intermediate to rounders.intermediate_form
