+++ src/rounders/overloads/float.py
@@ -6,7 +6,7 @@ import struct
 from typing import Optional, cast
 
 from rounders.generics import decade, is_finite, is_zero, preround, to_type_of
-from rounders.intermediate import IntermediateForm
+from rounders.intermediate_form import IntermediateForm
 
 #: Useful constants
 _TEN = fractions.Fraction(10)
