+++ src/rounders/overloads/fraction.py
@@ -4,7 +4,7 @@ import fractions
 from typing import Optional, cast
 
 from rounders.generics import decade, is_finite, is_zero, preround, to_type_of
-from rounders.intermediate import IntermediateForm
+from rounders.intermediate_form import IntermediateForm
 
 
 @decade.register
